+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Changes to EncryptionConfig in order to fully work with `encryption=provider-operator`.
+
 ## [10.13.0] - 2022-01-19
 
 ### Changed